// src/api/schemas/communityDetails.ts
/**
 * Generated by orval v6.29.1 🍺
 * Do not edit manually.
 * MyApp API
 * OpenAPI spec version: 1.0.0
 */
import type { CommunityDetailsBannerPicUrl } from './communityDetailsBannerPicUrl';
import type { CommunityDetailsCreatedAt } from './communityDetailsCreatedAt';
import type { CommunityDetailsJoinRequestStatus } from './communityDetailsJoinRequestStatus';
import type { CommunityDetailsProfilePicUrl } from './communityDetailsProfilePicUrl';

export interface CommunityDetails {
  banner_pic_url: CommunityDetailsBannerPicUrl;
  created_at: CommunityDetailsCreatedAt;
  description: string;
  id: number;
  is_admin?: boolean;
  is_member?: boolean;
  is_moderator?: boolean;
  is_reviewer?: boolean;
  join_request_status?: CommunityDetailsJoinRequestStatus;
  name: string;
  num_articles: number;
  num_members: number;
  num_moderators: number;
  num_published_articles: number;
  num_reviewers: number;
  profile_pic_url: CommunityDetailsProfilePicUrl;
  rules: string[];
  slug: string;
  tags: string;
  type: string;
}